Replace usage of DoL events for pet releases, fix Charm related issues * Fixes release button not working sometimes. Caused by effects being replaced by a new one when being refreshed, and DoL's event system losing the spell handler (since it uses weak references). * Fixes pulsing charms being allowed to add a new brain to the NPC when it shouldn't, resulting in a memory leak and ignoring the previous brain (if there was any). * Fixes Animist being unable to release Forestheart Ambusher. Fixed by allowing any pet to be released with the release command (Theurgist pets, turrets, and Bonedancer sub pets were already allowed, so there wasn't really any reason to not make a general rule out of it instead). * Reduces the amount of redundant work done in `ABrain.Stop` and overrides. It is sometimes called 3 or 4 times.

+++ GameServer/gameobjects/CharacterClasses/Albion/CharacterClassNecromancer.cs
@@ -46,8 +46,7 @@ namespace DOL.GS
 		/// </summary>
 		public override void CommandNpcRelease()
 		{
-			m_savedPetHealthPercent = (Player.ControlledBrain != null) ? (int)Player.ControlledBrain.Body.HealthPercent : 0;
-
+			m_savedPetHealthPercent = Player.ControlledBrain != null ? Player.ControlledBrain.Body.HealthPercent : 0;
 			base.CommandNpcRelease();
 			OnPetReleased();
 		}
